+++ src/app/lhc-form/lhc-form.component.ts
@@ -185,7 +185,7 @@ export class LhcFormComponent implements OnInit, OnChanges, OnDestroy, AfterView
                     })
                     .catch(error => {
                       const e = typeof error === "string" ? error : error.message
-                      self.onError.emit(e)
+                      self.onError.emit([e])
                     });
                 }
                 else {
@@ -193,8 +193,10 @@ export class LhcFormComponent implements OnInit, OnChanges, OnDestroy, AfterView
                 }
               })
               .catch(error => {
-                const e = typeof error === "string" ? error : error.message
-                self.onError.emit(e)
+                // loadFHIRResources fails with an array of errors, but perhaps
+                // something else has thrown an exception.
+                const e = error.message ? [error.message] : error;
+                self.onError.emit(e);
               });
             }
             else {
@@ -204,7 +206,7 @@ export class LhcFormComponent implements OnInit, OnChanges, OnDestroy, AfterView
           }
           catch (error) {
             const e = typeof error === "string" ? error : error.message
-            self.onError.emit(e)
+            self.onError.emit([e])
           };
         },1)
       }