+++ src/lib/lforms/lhc-form-utils.js
@@ -35,8 +35,8 @@ const FormUtils = {
    *  element itself.  The contents of this element will be replaced by the form.
    *  This element should be outside the scope of any existing AngularJS app on
    *  the page.
-   * @param {Object} [options] A hash of options. See avaialble options under templateOptions in
-   * form_definition.md. 'preppopulate' and 'fhirVersion' are not options in the templateOptions,
+   * @param {Object} [options] A hash of options. See available options under templateOptions in
+   * form_definition.md. 'prepopulate' and 'fhirVersion' are not options in the templateOptions,
    * but are included in the 'options' parameter.
    * @param {boolean} [options.prepopulate] Set to true if you want FHIR prepopulation to happen (if
    *  the form was an imported FHIR Questionnaire).
@@ -47,7 +47,8 @@ const FormUtils = {
    * lforms API calls when specified.
    * @return a Promise that will resolve after any needed external FHIR
    *  resources have been loaded (if the form was imported from a FHIR
-   *  Questionnaire).
+   *  Questionnaire).  If the needed resources fail, the promise will be
+   *  rejected and will contain an array of error messages.
    */
   addFormToPage: function(formDataDef, formContainer, options) {
 
@@ -94,7 +95,7 @@ const FormUtils = {
           resolve()
         });
         eleLhcForm.addEventListener('onError', function(e){
-          reject(e.detail)
+          reject(e.detail); // e.detail will be an array -- see lhc-form.component.ts
         });
       }
       catch(e) {